data/requirements.txt
python-dotenv
# openjdk-17-jre-headless
google-cloud-bigquery==3.11.4
google-cloud-core==2.3.3
google-cloud-storage
# google-cloud-dataproc
# pyspark==3.1.1
pandas
# findspark
numpy==1.20.1
requests
sqlalchemy==1.4.22
gcsfs
fsspec
seaborn
